+++ website_sale_loyalty_page/i18n/website_sale_loyalty_page.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 15.0\n"
+"Project-Id-Version: Odoo Server 16.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -19,7 +19,7 @@ msgid "<span>No promotions available</span>"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__can_publish
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__can_publish
 msgid "Can Publish"
 msgstr ""
 
@@ -29,42 +29,42 @@ msgid "Close"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model,name:website_sale_loyalty_page.model_coupon_program
-msgid "Loyalty display on a website"
-msgstr ""
-
-#. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_1920
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__image_1920
 msgid "Image"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_1024
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__image_1024
 msgid "Image 1024"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_128
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__image_128
 msgid "Image 128"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_256
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__image_256
 msgid "Image 256"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__image_512
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__image_512
 msgid "Image 512"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__is_published
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__is_published
 msgid "Is Published"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,help:website_sale_loyalty_page.field_coupon_program__public_name
+#: model:ir.model,name:website_sale_loyalty_page.model_loyalty_program
+msgid "Loyalty display on a website"
+msgstr ""
+
+#. module: website_sale_loyalty_page
+#: model:ir.model.fields,help:website_sale_loyalty_page.field_loyalty_program__public_name
 msgid "Name of the promo showed on website bellow the banner image."
 msgstr ""
 
@@ -74,26 +74,26 @@ msgid "Promotions"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__public_name
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__public_name
 msgid "Public Name"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,help:website_sale_loyalty_page.field_coupon_program__website_url
+#: model:ir.model.fields,help:website_sale_loyalty_page.field_loyalty_program__website_url
 msgid "The full URL to access the document through the website."
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__website_published
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__website_published
 msgid "Visible on current website"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model_terms:ir.ui.view,arch_db:website_sale_loyalty_page.coupon_program_view_form_common
+#: model_terms:ir.ui.view,arch_db:website_sale_loyalty_page.loyalty_program_view_form
 msgid "Website"
 msgstr ""
 
 #. module: website_sale_loyalty_page
-#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_coupon_program__website_url
+#: model:ir.model.fields,field_description:website_sale_loyalty_page.field_loyalty_program__website_url
 msgid "Website URL"
 msgstr ""
