+++ website_sale_loyalty_page/models/loyalty_program.py
@@ -4,10 +4,10 @@
 from odoo import fields, models
 
 
-class CouponProgram(models.Model):
-    _name = "coupon.program"
-    _inherit = ["coupon.program", "image.mixin", "website.published.mixin"]
-    _description = "Coupon display on a website"
+class LoyaltyProgram(models.Model):
+    _name = "loyalty.program"
+    _inherit = ["loyalty.program", "image.mixin", "website.published.mixin"]
+    _description = "Loyalty display on a website"
 
     public_name = fields.Html(
         help="Name of the promo showed on website bellow the banner image.",
