+++ website_sale_loyalty_page/readme/CONFIGURE.rst
@@ -1,9 +1,9 @@
 To show the promotions/coupons on website you need to:
 
-#. Go to *Sales > Products > Promotion Programs/Coupon Programs*.
+#. Go to *Website > eCommerce > Loyalty > Discount & Loyalty*.
 #. Select the program that we will publish.
-#. In the Website section, upload an image, add a text or both.
-#. Press on the smart button for publishing the promotion/coupon.
+#. In the Website tab, upload an image, add a text or both.
+#. Press on the smart button for publishing the loyalty.
 
 **Note:** The order of the items is defined by a sequence, but the promotions will
 appear before the coupons.
